# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoDoctor MVP API
  version: 1.2.0-draft
  description: |
    Контракт REST API guest-first manual-first MVP AutoDoctor по ТЗ v0.8.
    Публичный browse shell и создание гостевой сессии не требуют приглашения.
    VIN необязателен и принимается только в запросах; ответы содержат nullable vin_masked.
    Реализуемый срез профиля автомобиля не вызывает внешний VIN/AI provider.
    Maintenance v3 включает активные mileage forecast, condition observations,
    plan-item UI preferences и идемпотентный batch/upsert history answers.
    Источник by-pilot-baseline-2 — внутренняя редакционная методика AutoDoctor,
    не OEM/manual; опубликованные версии остаются immutable.
    /vehicles/decode и /vehicles/confirm сохранены только как будущий контракт enrichment.
    Все ресурсы автомобиля адресуются с vehicleId; сервер проверяет текущего владельца.
    Все даты и время передаются в UTC в формате RFC 3339.
    Filament AI admin использует внутренний серверный интерфейс: публичные мобильные
    endpoints конфигурации, playground и A/B experiments в этом контракте отсутствуют.
servers:
  - url: https://api-dev.autodoctor.by/api/v1
    description: Development
  - url: https://api.autodoctor.by/api/v1
    description: Production
...
security:
  - SessionAuth: []
  - BearerAuth: []
paths:
...
  /consents:
    post:
      tags: [Consents]
      operationId: grantConsents
      summary: Зафиксировать решения по согласиям
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantConsentsRequest"
      responses:
        "201":
          description: Решения сохранены
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConsentReceipt"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
  /vehicles/{vehicleId}/condition-observations:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Maintenance, Consumables]
      operationId: listConditionObservations
      summary: Получить наблюдения фактического износа
      description: |
        Append-only observations numeric MVP subset: brake_pads, brake_discs,
        tire_condition_inspection. Newest first; отсутствие записи не порождает fake %.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Страница наблюдений
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionObservationList"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Maintenance, Consumables]
      operationId: createConditionObservation
      summary: Добавить фактический процент износа
      description: |
        Идемпотентно добавляет новое observation и пересчитывает maintenance-v3 plan.
        wear_percent: 0=new/no wear, 100=fully worn; remaining_percent read-only.
        Install date не требуется. Latest observation feeds plan и consumables, но не
        создаёт OEM remaining-life claim и не вычисляет next due из wear.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConditionObservationCreate"
            example:
              work_code: brake_pads
              wear_percent: 76
              observed_at: "2026-07-20"
              mileage: {value: 84250, unit: km}
              source: self
              note: Визуальная оценка
      responses:
        "201":
          description: Наблюдение добавлено, план пересчитан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionObservationCreateResult"
              example:
                observation:
                  id: 6a17dc13-cb98-42d5-82a2-07be550d44df
                  vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                  work_code: brake_pads
                  wear_percent: 76
                  remaining_percent: 24
                  observed_at: "2026-07-20"
                  mileage: {value: 84250, unit: km}
                  source: self
                  note: Визуальная оценка
                  created_at: "2026-07-20T10:15:00Z"
                maintenance_plan_id: c3ecf7b6-eaab-4365-a9df-c1e34a822170
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /vehicles/{vehicleId}/plan-item-ui-preferences:
    x-future-contract: true
    description: |
      Dormant/reversible contract. Активные Laravel routes отсутствуют, mobile MVP
      не читает и не записывает preferences. Существующая таблица может сохраняться,
      но preference никогда не влияет на plan, comparator/order, warnings или notifications.
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [Maintenance]
      operationId: listPlanItemUiPreferences
      summary: Получить визуальные настройки пунктов плана
      description: |
        Future-only presentation overlay keyed by work_code. Не входит в active contract.
        Preferences не меняют plan/comparator/order/warnings/notifications и не входят
        в snapshot input_hash/content_hash.
      responses:
        "200":
          description: Все сохранённые preferences автомобиля
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItemUiPreferenceList"
              example:
                items:
                  - vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                    work_code: brake_pads
                    collapsed: true
                    version: 1
                    created_at: "2026-07-20T10:20:00Z"
                    updated_at: "2026-07-20T10:20:00Z"
        default:
          $ref: "#/components/responses/Error"
    put:
      tags: [Maintenance]
      operationId: putPlanItemUiPreference
      summary: Будущая настройка скрытия пункта плана
      description: |
        Idempotent upsert одного work_code. version=0 означает create; для существующей
        строки требуется текущая optimistic version, stale version возвращает 409.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemUiPreferencePut"
            example: {work_code: brake_pads, collapsed: true, version: 0}
      responses:
        "200":
          description: Preference сохранена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItemUiPreference"
              example:
                vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                work_code: brake_pads
                collapsed: true
                version: 1
                created_at: "2026-07-20T10:20:00Z"
                updated_at: "2026-07-20T10:20:00Z"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Токен анонимной гостевой сессии.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Короткоживущий access-токен зарегистрированного пользователя.
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        Язык локализованных текстов и ошибок: `ru` или `en`; региональные теги
        (`ru-BY`, `en-US`) нормализуются до базового языка. Для публичных запросов
        и ошибок до аутентификации отсутствие или неподдерживаемое значение даёт `ru`.
        После успешной аутентификации анонимной сессии отсутствие заголовка использует
        сохранённый `session.locale`; переданный заголовок имеет приоритет.
      schema:
        type: string
        examples: [ru, en, ru-BY, en-US]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Уникальный UUID операции; повтор с тем же телом возвращает прежний результат.
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Error:
      description: Ошибка запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VEHICLE_NOT_SUPPORTED
              message: Для этого автомобиля проверенные данные пока не подготовлены.
              request_id: req_01J2Y6D8QK3R7T5V9W1X
              details:
                support_request_id: 7ea2c18f-a6cf-4c19-95c4-c4cb56a701b7
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_FAILED
              message: Проверьте заполненные поля.
              request_id: req_01J2Y6D8QK3R7T5V9W1X
              fields:
                mileage.value:
                  - Значение должно быть неотрицательным.
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              pattern: "^[A-Z][A-Z0-9_]+$"
            message:
              type: string
            request_id:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object
              additionalProperties: true
...
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
    MileageUnit:
      type: string
      enum: [km, mi]
    Mileage:
      type: object
      additionalProperties: false
      required: [value, unit]
      properties:
        value:
          type: integer
          minimum: 0
        unit:
          $ref: "#/components/schemas/MileageUnit"
...
    ConsentPurpose:
      type: string
      enum: [essential_processing, ai_processing, analytics, marketing]
...
    GrantConsentsRequest:
      type: object
      additionalProperties: false
      required: [decisions]
      properties:
        decisions:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [purpose, version, granted]
            properties:
              purpose:
                $ref: "#/components/schemas/ConsentPurpose"
              version:
                type: string
              granted:
                type: boolean
    ConsentReceipt:
      type: object
      additionalProperties: false
      required: [id, purpose, version, granted, decided_at]
      properties:
        id:
          type: string
          format: uuid
        purpose:
          $ref: "#/components/schemas/ConsentPurpose"
        version:
          type: string
        granted:
          type: boolean
        decided_at:
          type: string
          format: date-time
...
    ServiceWorkCode:
      type: string
      enum: [engine_oil, oil_filter, air_filter, cabin_filter, brake_fluid, coolant, brake_system_inspection, tire_condition_inspection, coolant_inspection, spark_plugs, timing_drive, transmission_oil, brake_pads, brake_discs, other]
...
    ConditionObservationSource:
      type: string
      enum: [self, workshop]
    ConditionWorkCode:
      type: string
      enum: [brake_pads, brake_discs, tire_condition_inspection]
...
    ConditionObservationCreate:
      type: object
      additionalProperties: false
      required: [work_code, wear_percent, observed_at, source]
      properties:
        work_code:
          $ref: "#/components/schemas/ConditionWorkCode"
        wear_percent:
          type: integer
          minimum: 0
          maximum: 100
          description: 0 = новая/без износа; 100 = полностью изношена.
        observed_at:
          type: string
          format: date
        mileage:
          oneOf:
            - $ref: "#/components/schemas/Mileage"
            - type: "null"
        source:
          $ref: "#/components/schemas/ConditionObservationSource"
        note:
          type: [string, "null"]
          maxLength: 2000
    ConditionObservation:
      type: object
      additionalProperties: false
      required: [id, vehicle_id, work_code, wear_percent, remaining_percent, observed_at, mileage, source, note, created_at]
      properties:
        id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        work_code:
          $ref: "#/components/schemas/ConditionWorkCode"
        wear_percent:
          type: integer
          minimum: 0
          maximum: 100
        remaining_percent:
          type: integer
          minimum: 0
          maximum: 100
          readOnly: true
          description: Всегда 100 - wear_percent; клиент не передаёт.
        observed_at:
          type: string
          format: date
        mileage:
          oneOf:
            - $ref: "#/components/schemas/Mileage"
            - type: "null"
        source:
          $ref: "#/components/schemas/ConditionObservationSource"
        note:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
    ConditionObservationCreateResult:
      type: object
      additionalProperties: false
      required: [observation, maintenance_plan_id]
      properties:
        observation:
          $ref: "#/components/schemas/ConditionObservation"
        maintenance_plan_id:
          type: string
          format: uuid
    ConditionObservationList:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConditionObservation"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    PlanItemUiPreferencePut:
      x-future-contract: true
      type: object
      additionalProperties: false
      required: [work_code, collapsed, version]
      properties:
        work_code:
          $ref: "#/components/schemas/ServiceWorkCode"
        collapsed:
          type: boolean
        version:
          type: integer
          minimum: 0
          description: 0 для create; текущая persisted version для update.
    PlanItemUiPreference:
      x-future-contract: true
      type: object
      additionalProperties: false
      required: [vehicle_id, work_code, collapsed, version, created_at, updated_at]
      properties:
        vehicle_id:
          type: string
          format: uuid
        work_code:
          $ref: "#/components/schemas/ServiceWorkCode"
        collapsed:
          type: boolean
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlanItemUiPreferenceList:
      x-future-contract: true
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlanItemUiPreference"
...